requirements.txt
numpy == 1.21.2
pandas == 1.3.5
python_dotenv == 0.20.0
requests == 2.28.1
tqdm == 4.62.2
